Ski_Resort_SQL_Database/reference/requirements.txt
# Python Dependencies for Event Ticketing System
# Install with: pip install -r requirements.txt

# MySQL Database Connector
mysql-connector-python>=8.0.0

# Table formatting for output
tabulate>=0.9.0

